services/gateway/pyproject.toml
[project]
name = "eci-gateway"
version = "0.1.0"
description = "ECI API Gateway"
requires-python = ">=3.11"
authors = [{name = "ECI Team", email = "team@example.com"}]
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.0",
  "httpx==0.27.0",
  "pydantic==2.9.2",
  "PyJWT==2.9.0",
  "strawberry-graphql==0.227.3",
  "redis==5.0.1",
  "cachetools==5.3.3",
  "python-multipart==0.0.9",
  "pydantic-settings==2.5.2"
]
